requirements.txt
gradio
gradiologin
requests
langchain-community
langchain-core
langchain-google-genai
langchain-groq
langchain-tavily
langgraph
python-dotenv
pandas
arxiv
wikipedia
pymupdf
google-genai
tavily-python